Change to using a uint32 for regions on sampling api
Why? Because Unity sucks, and because we can afford to upgrade later if necessary. YAGNI.

+++ src/BodyView.h
@@ -13,7 +13,7 @@ struct BodyView {
 
 
 
-	int getNodeRegion(uint32_t nodeIndex, uint64_t * outRegion)
+	int getNodeRegion(uint32_t nodeIndex, uint32_t * outRegion)
 	{
 		*outRegion = pairs[nodeIndex].Region;
 		return NSVR_Success_Unqualified;
